issue solved

+++ src/main/java/unicam/repository/InMemoryConsegnaRepo.java
@@ -7,10 +7,10 @@ import java.util.Map;
 
 public class InMemoryConsegnaRepo implements ConsegnaRepository {
 
-    private Map<Integer, Consegna> consegne;
+    private Map<Long, Consegna> consegne;
 
     public InMemoryConsegnaRepo() {
-        consegne = new HashMap<Integer, Consegna>();
+        consegne = new HashMap<Long, Consegna>();
     }
 
     @Override
